Fix lint and move widgets to dedicated directory
